+++ src/ts/ui/index.ts
@@ -1,11 +1,11 @@
 // ui is massive fuckin w
-
 export * as accordion from './accordion';
 export * as header from './header';
 export * as info from './info';
 export * as input from './input';
 export * as intro from './intro';
 export * as loading from './loading';
+export * as select from './select';
 export * as settings from './settings';
 export * as stats from './stats';
 export * as theme from './theme';